move auto-sharding to xla/service/spmd folder
