+++ tensorflow/compiler/xla/service/gpu/BUILD
@@ -1412,7 +1412,6 @@ cc_library(
     ],
     deps = [
         ":alias_passthrough_params",
-        ":auto_sharding",
         ":cudnn_batchnorm_rewriter",
         ":fusion_bitcast_lift",
         ":fusion_merger",
@@ -1516,6 +1515,7 @@ cc_library(
         "//tensorflow/compiler/xla/service:zero_sized_hlo_elimination",
         "//tensorflow/compiler/xla/service/gpu/llvm_gpu_backend",
         "//tensorflow/compiler/xla/service/llvm_ir:llvm_util",
+        "//tensorflow/compiler/xla/service/spmd:auto_sharding",
         "//tensorflow/core:lib",
         "//tensorflow/core:lib_internal",
         "//tensorflow/core/platform:regexp",
@@ -2045,23 +2045,6 @@ cc_library(
         "//tensorflow/compiler/xla:shape_util",
         "//tensorflow/compiler/xla/service:hlo",
         "//tensorflow/compiler/xla/service:hlo_pass",
-    ],
-)
-
-cc_library(
-    name = "auto_sharding",
-    srcs = ["auto_sharding.cc"],
-    hdrs = ["auto_sharding.h", "auto_sharding_util.h"],
-    deps = [
-        "//tensorflow/compiler/xla:shape_util",
-        "//tensorflow/compiler/xla/service:dump",
-        "//tensorflow/compiler/xla/service:hlo",
-        "//tensorflow/compiler/xla/service:hlo_live_range",
-        "//tensorflow/compiler/xla/service:hlo_memory_scheduler",
-        "//tensorflow/compiler/xla/service:hlo_pass",
-        "//tensorflow/compiler/xla/service:hlo_sharding_util",
-        "//tensorflow/compiler/xla/service:pass_context",
-        "@pybind11",
     ],
 )
