+++ tensorflow/compiler/xla/service/spmd/BUILD
@@ -148,3 +148,20 @@ cc_library(
         "@com_google_absl//absl/container:flat_hash_map",
     ],
 )
+
+cc_library(
+    name = "auto_sharding",
+    srcs = ["auto_sharding.cc"],
+    hdrs = ["auto_sharding.h", "auto_sharding_util.h"],
+    deps = [
+        "//tensorflow/compiler/xla:shape_util",
+        "//tensorflow/compiler/xla/service:dump",
+        "//tensorflow/compiler/xla/service:hlo",
+        "//tensorflow/compiler/xla/service:hlo_live_range",
+        "//tensorflow/compiler/xla/service:hlo_memory_scheduler",
+        "//tensorflow/compiler/xla/service:hlo_pass",
+        "//tensorflow/compiler/xla/service:hlo_sharding_util",
+        "//tensorflow/compiler/xla/service:pass_context",
+        "@pybind11",
+    ],
+)
